+++ src/checks.jl
@@ -35,7 +35,57 @@ function EMB.check_node(
 end
 
 """
-    EMB.check_node(n::CSPandPV, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
+    EMB.check_node(n::MultipleBuildingTypes, 𝒯, ::EnergyModel, ::Bool)
+
+This method checks that the [`MultipleBuildingTypes`](@ref) node is valid.
+
+## Checks
+- The field `cap_p` is required to be non-negative for all resources `p`.
+- The values of the dictionary `input` are required to be non-negative.
+- The sum of the fields `penalty_surplus` and `penalty_deficit` has to be
+  non-negative to avoid an infeasible model.
+"""
+function EMB.check_node(n::MultipleBuildingTypes, 𝒯, ::EnergyModel, ::Bool)
+    𝒫 = inputs(n)
+    for p ∈ 𝒫
+        @assert_or_log(
+            isempty(setdiff([p], keys(EMB.capacity(n)))),
+            "Resource $(p) is not included in the dictionary corresponding to the field `cap`"
+        )
+        @assert_or_log(
+            isempty(setdiff([p], keys(EMB.surplus_penalty(n)))),
+            "Resource $(p) is not included in the dictionary corresponding to the field `penalty_surplus`"
+        )
+        @assert_or_log(
+            isempty(setdiff([p], keys(EMB.deficit_penalty(n)))),
+            "Resource $(p) is not included in the dictionary corresponding to the field `penalty_deficit`"
+        )
+    end
+    if isempty(setdiff(𝒫, keys(EMB.capacity(n))))
+        @assert_or_log(
+            all(EMB.capacity(n, t, p) ≥ 0 for t ∈ 𝒯, p ∈ 𝒫),
+            "The capacity must be non-negative."
+        )
+    end
+    @assert_or_log(
+        all(inputs(n, p) ≥ 0 for p ∈ 𝒫),
+        "The values for the Dictionary `input` must be non-negative."
+    )
+
+    if isempty(setdiff(𝒫, keys(EMB.surplus_penalty(n)))) &&
+       isempty(setdiff(𝒫, keys(EMB.deficit_penalty(n))))
+        @assert_or_log(
+            all(
+                EMB.surplus_penalty(n, t, p) + EMB.deficit_penalty(n, t, p) ≥ 0 for t ∈ 𝒯,
+                p ∈ 𝒫
+            ),
+            "An inconsistent combination of `penalty_surplus` and `penalty_deficit` leads to an infeasible model."
+        )
+    end
+end
+
+"""
+    EMB.check_node(n::CSPandPV, 𝒯, ::EnergyModel, check_timeprofiles::Bool)
 
 This method checks that the *[`CSPandPV`](@ref)* node is valid.
 
@@ -49,15 +99,35 @@ This method checks that the *[`CSPandPV`](@ref)* node is valid.
 - The profiles in `opex_fixed_p` have to have the same length as the number of strategic
   periods.
 """
-function EMB.check_node(n::CSPandPV, 𝒯, modeltype::EnergyModel, check_timeprofiles::Bool)
+function EMB.check_node(n::CSPandPV, 𝒯, ::EnergyModel, check_timeprofiles::Bool)
     𝒯ᴵⁿᵛ = strategic_periods(𝒯)
 
     𝒫 = outputs(n)
+    for p ∈ 𝒫
+        @assert_or_log(
+            isempty(setdiff([p], keys(EMB.capacity(n)))),
+            "Resource $(p) is not included in the dictionary corresponding to the field `cap`"
+        )
+        @assert_or_log(
+            isempty(setdiff([p], keys(EMB.opex_fixed(n)))),
+            "Resource $(p) is not included in the dictionary corresponding to the field `opex_fixed`"
+        )
+        @assert_or_log(
+            isempty(setdiff([p], keys(EMB.opex_var(n)))),
+            "Resource $(p) is not included in the dictionary corresponding to the field `opex_var`"
+        )
+        @assert_or_log(
+            isempty(setdiff([p], keys(EMR.profile(n)))),
+            "Resource $(p) is not included in the dictionary corresponding to the field `profile`"
+        )
+    end
 
-    @assert_or_log(
-        all(EMB.capacity(n, t, p) ≥ 0 for t ∈ 𝒯, p ∈ 𝒫),
-        "The capacity must be non-negative."
-    )
+    if isempty(setdiff(𝒫, keys(EMB.capacity(n))))
+        @assert_or_log(
+            all(EMB.capacity(n, t, p) ≥ 0 for t ∈ 𝒯, p ∈ 𝒫),
+            "The capacity must be non-negative."
+        )
+    end
     @assert_or_log(
         all(outputs(n, p) ≥ 0 for p ∈ 𝒫),
         "The values for the Dictionary `output` must be non-negative."
@@ -86,12 +156,14 @@ function EMB.check_node(n::CSPandPV, 𝒯, modeltype::EnergyModel, check_timepro
         )
     end
 
-    @assert_or_log(
-        all(EMR.profile(n, t, p) ≤ 1 for t ∈ 𝒯, p ∈ 𝒫),
-        "The profile field must be less or equal to 1."
-    )
-    @assert_or_log(
-        all(EMR.profile(n, t, p) ≥ 0 for t ∈ 𝒯, p ∈ 𝒫),
-        "The profile field must be non-negative."
-    )
+    if isempty(setdiff(𝒫, keys(EMR.profile(n))))
+        @assert_or_log(
+            all(EMR.profile(n, t, p) ≤ 1 for t ∈ 𝒯, p ∈ 𝒫),
+            "The profile field must be less or equal to 1."
+        )
+        @assert_or_log(
+            all(EMR.profile(n, t, p) ≥ 0 for t ∈ 𝒯, p ∈ 𝒫),
+            "The profile field must be non-negative."
+        )
+    end
 end