+++ src/model.jl
@@ -1,3 +1,18 @@
+"""
+    EMB.variables_node(m, 𝒩::Vector{MultipleBuildingTypes}, 𝒯, ::EnergyModel)
+
+For a [`MultipleBuildingTypes`](@ref) node, the following variables are created:
+- `buildings_surplus[n, t, p]` is the surplus of node `n` with resource `p` in operational period `t`.
+- `buildings_deficit[n, t, p]` is the surplus of node `n` with resource `p` in operational period `t`.
+"""
+function EMB.variables_node(m, 𝒩::Vector{MultipleBuildingTypes}, 𝒯, ::EnergyModel)
+
+    # Declaration of the required subsets.
+    𝒫 = unique([p for n ∈ 𝒩 for p ∈ inputs(n)])
+
+    @variable(m, buildings_surplus[𝒩, 𝒯, 𝒫] >= 0)
+    @variable(m, buildings_deficit[𝒩, 𝒯, 𝒫] >= 0)
+end
 
 """
     EMB.variables_node(m, 𝒩::Vector{CSPandPV}, 𝒯, ::EnergyModel)